refactor: Board 네이밍 변경
