+++ src/main/java/chess/model/piece/Empty.java
@@ -1,6 +1,6 @@
 package chess.model.piece;
 
-import chess.model.board.ConsoleBoard;
+import chess.model.board.Board;
 import chess.model.square.Direction;
 import chess.model.square.Square;
 
@@ -29,7 +29,7 @@ public class Empty extends Piece {
     }
 
     @Override
-    public boolean canMoveWithoutObstacle(ConsoleBoard consoleBoard, Square source, Square target) {
+    public boolean canMoveWithoutObstacle(Board board, Square source, Square target) {
         return false;
     }
