+++ src/main/java/chess/model/piece/PointMovingPiece.java
@@ -1,6 +1,6 @@
 package chess.model.piece;
 
-import chess.model.board.ConsoleBoard;
+import chess.model.board.Board;
 import chess.model.square.Square;
 
 import java.util.Collections;
@@ -22,8 +22,8 @@ public abstract class PointMovingPiece extends Piece {
                 .anyMatch(direction -> source.findLocation(direction, target));
     }
 
-    public boolean canMoveWithoutObstacle(ConsoleBoard consoleBoard, Square source, Square target) {
-        Piece targetPiece = consoleBoard.get(target);
+    public boolean canMoveWithoutObstacle(Board board, Square source, Square target) {
+        Piece targetPiece = board.get(target);
         return isNotAlly(targetPiece);
     }
