+++ src/test/java/chess/model/piece/BishopTest.java
@@ -1,6 +1,6 @@
 package chess.model.piece;
 
-import chess.model.board.ConsoleBoard;
+import chess.model.board.Board;
 import chess.model.square.File;
 import chess.model.square.Rank;
 import chess.model.square.Square;
@@ -13,11 +13,11 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 public class BishopTest {
 
-    private ConsoleBoard consoleBoard;
+    private Board board;
 
     @BeforeEach
     public void setUp() {
-        consoleBoard = new ConsoleBoard();
+        board = new Board();
     }
 
     @Test
@@ -51,7 +51,7 @@ public class BishopTest {
         Square source = Square.of(File.B, Rank.EIGHT);
         Square target = Square.of(File.A, Rank.SEVEN);
 
-        assertThat(bishop.canMoveWithoutObstacle(consoleBoard, source, target)).isFalse();
+        assertThat(bishop.canMoveWithoutObstacle(board, source, target)).isFalse();
     }
 
     @Test
@@ -60,6 +60,6 @@ public class BishopTest {
         Square source = Square.of(File.B, Rank.EIGHT);
         Square target = Square.of(File.D, Rank.SIX);
 
-        assertThat(bishop.canMoveWithoutObstacle(consoleBoard, source, target)).isFalse();
+        assertThat(bishop.canMoveWithoutObstacle(board, source, target)).isFalse();
     }
 }
