+++ src/test/java/chess/model/piece/PawnTest.java
@@ -1,6 +1,6 @@
 package chess.model.piece;
 
-import chess.model.board.ConsoleBoard;
+import chess.model.board.Board;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -10,11 +10,11 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 class PawnTest {
 
-    private ConsoleBoard consoleBoard;
+    private Board board;
 
     @BeforeEach
     public void setUp() {
-        consoleBoard = new ConsoleBoard();
+        board = new Board();
     }
 
     @Test
@@ -30,8 +30,8 @@ class PawnTest {
         Pawn pawn = new Pawn(Team.WHITE);
 
         assertAll(
-                () -> assertThat(pawn.movable(consoleBoard, E2, E4)).isTrue(),
-                () -> assertThat(pawn.movable(consoleBoard, C2, C3)).isTrue());
+                () -> assertThat(pawn.movable(board, E2, E4)).isTrue(),
+                () -> assertThat(pawn.movable(board, C2, C3)).isTrue());
     }
 
     @Test
@@ -39,14 +39,14 @@ class PawnTest {
         Pawn pawn = new Pawn(Team.WHITE);
 
         assertAll(
-                () -> assertThat(pawn.movable(consoleBoard, E2, E5)).isFalse(),
-                () -> assertThat(pawn.movable(consoleBoard, E2, D3)).isFalse());
+                () -> assertThat(pawn.movable(board, E2, E5)).isFalse(),
+                () -> assertThat(pawn.movable(board, E2, D3)).isFalse());
     }
 
     @Test
     void WhitePawnCannotMoveInLastColumn() {
         Pawn pawn = new Pawn(Team.WHITE);
 
-        assertThat(pawn.movable(consoleBoard, E2, E1)).isFalse();
+        assertThat(pawn.movable(board, E2, E1)).isFalse();
     }
 }
